refacto(discord-bot): modularize utils
